Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,12 +1,12 @@
 // Your code here...
 #include <stdio.h>
 int main(){
-    int num1;
-    int num2;
-    int num3;
+    float num1;
+    float num2;
+    float num3;
 
 
-    scanf("%i %i %i",&num1,&num2,&num3);
+    scanf("%f %f %f",&num1,&num2,&num3);
 
     float average = (num1+num2+num3)/3;
     printf("Average: %.2f",average);
